Changes to multilingual fallback
Increased and decreased coverage for labels and descriptions respectively:

For labels: If the item has no label in our prioritized list of languages we
explicitly query for (which are based on the larger Wikipedias), but it
does have a label in SOME language, fall back to randomly selecting one of
those labels, to avoid falling back to Qxxx numeric labels.

For descriptions: On the other hand, for descriptions, use only English labels
with no fallback. The rest of the interface is currently monolingual, so it's
weird to sometimes have non-English descriptions in parentheses in a narrative
stream that otherwise doesn't support localization.