Prioritize specific HQ coordinates
If there is a qualifier P625 (coordinate location) on the P159 (headquarters
location) statement, use *only* those coordinates, and don't consider the
coordinates of the city/place that's the target of the P159 statement. This
was the intent of the change in the previous commit, but that commit actually
treated both sets of coordinates equally. The goal is to use specific HQ
coordinates when available, to reduce how many organizations get bunched up
around the "official" coordinates of a big city like Houston or London.